requirements.txt
Appium-Python-Client==4.5.0
selenium==4.27.1
python-dotenv==1.0.1
requests==2.31.0